+++ lib/src/renderer/get_pixels/main.dart
@@ -3,12 +3,7 @@ import 'dart:typed_data';
 import 'io.dart';
 
 Future<Uint8List> getPixels({
-  String? path,
-  List<int>? bytes,
+  required String path,
   bool removeTempFile = true,
 }) =>
-    getPlatformPixels(
-      path: path,
-      bytes: bytes,
-      removeTempFile: removeTempFile,
-    );
+    getPlatformPixels(path: path, removeTempFile: removeTempFile);